moved code to series from solvers, added tests and initial conditions
