+++ sympy/series/tests/test_series.py
@@ -213,3 +213,33 @@ def test_issue_15539():
         + O(x**(-6), (x, -oo)))
     assert series(atan(x), x, oo) == (-1/(5*x**5) + 1/(3*x**3) - 1/x + pi/2
         + O(x**(-6), (x, oo)))
+
+
+def test_lagrange_inversion_theorem():
+    from sympy.series.series import lagrange_inversion_theorem
+    from sympy.abc import x, y
+    from sympy import exp
+
+    eq = 2 * y
+    raises(AssertionError, lambda: lagrange_inversion_theorem(eq))
+
+    eq1 = 2
+    raises(AssertionError, lambda: lagrange_inversion_theorem(eq1))
+
+    eq = x ** 2
+    assert lagrange_inversion_theorem(eq) == x**2/2 + (x**2 - 1)**3/16 - (x**2 - 1)**2/8 - 1/2
+    assert lagrange_inversion_theorem(eq, 1) == x**2/2 - 1/2
+    assert lagrange_inversion_theorem(eq, 2) == x**2/2 - (x**2 - 1)**2/8 - 1/2
+    assert lagrange_inversion_theorem(eq, 4) == x**2/2 - 5*(x**2 - 1)**4/128 + (x**2 - 1)**3/16 - (x**2 - 1)**2/8 - 1/2
+
+    eq = (x ** 3) * exp(x)
+    assert lagrange_inversion_theorem(eq, 1) == (x**3*exp(x) - E)*exp(-1)/4
+
+    eq = (x ** 3) * exp(x) * y
+    raises(AssertionError, lambda: lagrange_inversion_theorem(eq, 1))
+
+    eq = sin(x)
+    assert lagrange_inversion_theorem(eq) == sin(x)**3/6 + sin(x)
+
+    eq = cos(y)
+    raises(AssertionError, lambda: lagrange_inversion_theorem(eq, 1))
