+++ sympy/solvers/solvers.py
@@ -22,7 +22,7 @@ from sympy.core import (S, Add, Symbol, Equality, Dummy, Expr, Mul,
 from sympy.core.exprtools import factor_terms
 from sympy.core.function import (expand_mul, expand_multinomial, expand_log,
                           Derivative, AppliedUndef, UndefinedFunction, nfloat,
-                          Function, expand_power_exp, Lambda, _mexpand, diff)
+                          Function, expand_power_exp, Lambda, _mexpand)
 from sympy.integrals.integrals import Integral
 from sympy.core.numbers import ilcm, Float
 from sympy.core.relational import Relational, Ge, _canonical
@@ -56,8 +56,6 @@ from sympy.solvers.inequalities import reduce_inequalities
 from types import GeneratorType
 from collections import defaultdict
 
-from sympy.series.limits import limit
-from sympy.functions.combinatorial.factorials import factorial
 
 import warnings
 
@@ -3141,118 +3139,6 @@ def _invert(eq, *symbols, **kwargs):
         if lhs == was:
             break
     return rhs, lhs
-
-def lagrange_inversion_theorem(eq, no_of_terms = 3):
-    """
-    The Lagrange inversion theorem (or Lagrange inversion formula, which we abbreviate
-    as LIT), also known as the Lagrange--Bürmann formula, gives the Taylor series
-    expansion of the inverse function of an analytic function.
-
-    The theorem was proved by Lagrange (1736--1813) and generalized by the German
-    mathematician and teacher Hans Heinrich Bürmann ( --1817), both in the late 18th
-    century.
-
-    The Lagrange inversion formula is one of the fundamental formulas of combinatorics.
-    In its simplest form it gives a formula for the power series coefficients of the
-    solution f(x) of the function equation f(x) = xG(f(x)) in terms of coefficients of
-    powers of G.
-
-    Theorem: Suppose z is defined as a function of w by an equation of the form
-
-    f(w) = z
-
-    where f is analytic at a point a and f'(a) is not equal to 0.
-    Then it is possible to invert or solve the equation for w in the form of a series:
-
-    w = a + \sum_{n = 1}^{oo} (g_n*(z - f(a))**n)/n!
-
-    where
-                      n-1
-                    d            w - a       n
-    g_n =  lim   ( ------- (  -----------  )   )
-          w --> a     n-1     f(w) - f(a)
-                   dw
-
-    The theorem further states that this series has a non-zero radius of convergence,
-    i.e., w represents an analytic function of z in a neighbourhood of z = f(a).
-    This is also called reversion of series.
-
-    Parameters
-    ==========
-
-    eq : expression, the expansion of whose inverse is to be found.
-
-    no_of_terms : no of terms required in the expansion; default value is 3 terms.
-
-    Examples
-    ========
-
-    >>> from sympy.solvers.solvers import lagrange_inversion_theorem
-    >>> from sympy import exp
-    >>> from sympy.abc import x
-    >>> eq = (x**2)*exp(x)
-
-    If no second argument is provided, the default number of terms i.e. 3, is present
-    in the expansion of the inverse of the provided equation or expression (as the first argument).
-
-    >>> lagrange_inversion_theorem(eq)
-    2*(x**2*exp(x) - E)**3*exp(-3)/27 - 7*(x**2*exp(x) - E)**2*exp(-2)/54 + (x**2*exp(x) - E)*exp(-1)/3
-
-    Here the number of terms required in the expansion is provided.
-
-    >>> lagrange_inversion_theorem(eq, 1)
-    (x**2*exp(x) - E)*exp(-1)/3
-    >>> lagrange_inversion_theorem(eq, 2)
-    -7*(x**2*exp(x) - E)**2*exp(-2)/54 + (x**2*exp(x) - E)*exp(-1)/3
-
-    >>> eq = exp(x)
-    >>> lagrange_inversion_theorem(eq, 2)
-    -(exp(x) - 1)**2/2 + exp(x) - 1
-    >>> lagrange_inversion_theorem(eq, 3)
-    (exp(x) - 1)**3/3 - (exp(x) - 1)**2/2 + exp(x) - 1
-
-    Note
-    ====
-
-    There is a straightforward derivation using complex analysis and contour integration;
-    the complex formal power series version is a consequence of knowing the formula for polynomials,
-    so the theory of analytic functions may be applied. Actually, the machinery from analytic
-    function theory enters only in a formal way in this proof, in that what is really needed is
-    some property of the formal residue, and a more direct formal proof is available.
-
-    References
-    ==========
-    .. [1] https://en.wikipedia.org/wiki/Lagrange_inversion_theorem
-    .. [2] http://mathworld.wolfram.com/LagrangeInversionTheorem.html
-    .. [3] http://www.cfm.brown.edu/people/dobrush/am33/Mathematica/lit.html
-    """
-    from sympy.abc import w, x
-
-    n = no_of_terms
-    eq = eq.subs(x, w)
-    z = eq
-
-    z_diff = diff(eq, w)
-    t = 0
-
-    while (True):
-        if z_diff.subs(w, t) == 0 or not (z_diff.subs(w, t)).is_real:
-            t = t + 1
-            continue
-        else:
-            break
-
-    a = t
-    f_a = z.subs(w, a)
-    sum = 0
-
-    while (n > 0):
-        l = limit(diff(((w - a) / (z - f_a)) ** n, w, n - 1), w, a)
-        sum = sum + (((z - f_a) ** n) / factorial(n)) * l
-        n = n - 1
-
-    sum = sum.subs(w, x)
-    return sum
 
 def unrad(eq, *syms, **flags):
     """ Remove radicals with symbolic arguments and return (eq, cov),
